# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Documentação completa da minha API Laravel",
        "contact": {
            "name": "Suporte",
            "url": "https://example.com/suporte",
            "email": "suporte@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local de desenvolvimento"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Servidor de produção"
        }
    ],
    "paths": {
...
        "/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Listar todos os quartos",
                "operationId": "3e38452d1f1cbd72dea1eb1f8b4eacc9",
                "responses": {
                    "200": {
                        "description": "Lista de quartos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Criar um novo quarto",
                "operationId": "db9e87557265233cd3de0b99455ac90f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "hotel_id",
                                    "hotel_name",
                                    "inventory_count",
                                    "name"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hotel_name": {
                                        "type": "string",
                                        "example": "Hotel Central"
                                    },
                                    "inventory_count": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Quarto Duplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quarto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar quarto"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/rooms/{room}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Exibir um quarto específico",
                "operationId": "7db0b664f2034dc6b3a75457173cfc57",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "ID do quarto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quarto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quarto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Atualizar um quarto",
                "operationId": "059b593b9ad04b056e143c20d18a1ce6",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "ID do quarto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hotel_name": {
                                        "type": "string",
                                        "example": "Hotel Central"
                                    },
                                    "inventory_count": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 10
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Quarto Duplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quarto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar quarto"
                    },
                    "404": {
                        "description": "Quarto não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Remover um quarto",
                "operationId": "1945f3d92d67c1093bd320b194b60ce0",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "ID do quarto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quarto removido com sucesso"
                    },
                    "400": {
                        "description": "Erro ao remover quarto"
                    },
                    "404": {
                        "description": "Quarto não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Room": {
                "title": "Room",
                "required": [
                    "id",
                    "hotel_id",
                    "hotel_name",
                    "inventory_count",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "hotel_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "hotel_name": {
                        "type": "string",
                        "example": "Hotel Central"
                    },
                    "inventory_count": {
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "type": "string",
                        "example": "Quarto Duplo"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}